Support RAML 1.0
This mostly consisted of rewriting tests to output the new format
that the raml-1-parser outputs, which seems massively more complex
than the old parser. Almost everything that is returned is now a
function which you have to call to get the data you actually want,
which seems weird.

Hopefully this doesn't break anything, but lets push this as a new
major release for semver's sake.